Harden production security defaults and tenant isolation (#79)

+++ src/models/account.rs
@@ -107,5 +107,5 @@ pub struct AccountBalancePerDay {
     pub account_id: Uuid,
     pub account_name: String,
     pub date: String,
-    pub balance: i32,
+    pub balance: i64,
 }
